Modifications to make Paris behave more like Kohana's ORM.
I wanted Paris models to use proper PHP constructors... This
required some significant refactoring of the inheritance chain.
Part of this was my desire to make

$user = new User;
$user->foo = "bar";
$user->save();

create a new user, much like Kohana 2.4's ORM.  (I wasn't sure
how to use Paris to properly insert rows, it wasn't intuitive and I
was dropping back to IdiORM.)  This feature was not as hard as I would have thought
and basically involves making Paris models subclass all the way from
ORM instances, and as soon as they are instantiated they have
individual "state".  So, the class is no longer a wrapper for a factory
in ORM_Wrapper.  I did not like the use of factories, I prefer everything
to be available to PHP "new" whenever plausible, and the factory should
just be a wrapper over that so it chains nicely.  (Kohana bias.)

The upshot is you actually get to kill a lot methods in class Model that
were just calling ORM's, because now they simply inherit.

Incidentally, I also added support for transactions via:
Model::start_transaction()
Model::commit()
Model::rollback()

Needed it for something, was straightforward to add.

One of my goals here was to make this work with vanilla IdiORM, which
I did not need to edit to make my changes to Paris.  So, hopefully if
IdiORM is updated with new query magic, it will just work here too.

I need to update the docs.